Add line highlighting
Users can now click on line numbers to highlight a line.

Line highlighting works by storing the file number and line number in
the hash of the current page's URL.

Each time the hash changes, the line will be highlighted and the line
will be scrolled into view except if the user sets the hash by clicking
on a line number.

Example: #1L6
File 1, line 6 will be highlighted