Add file drop functionality
As a user with many files to upload, it's probably quicker to just drag
and drop from a file explorer (or something to that effect), however
said functionality doesn't exist yet.

This commit implements the file drag and drop functionality,
here we abuse the `autodetect` lexer instead of parsing the file
extension (or content) and picking the "correct" lexer which is
somewhat harder.