Move file upload section down
This is not the most important input method, and as such shouldn't be on
top, this should be viewable on most browser by default even at the
bottom.

Also add a `margin-bottom` to `paste-submit` for spacing consistency,
since the upload section is below the paste submit section, without
this, the two sections will look like they are merged into one.