pinnwand: use aware datetime objects
Deprecation warnings are being emitted about using `datetime.utcnow` and
to replace those with `datetime.now(timezone.utc)`. This creates a
timezone aware datetime object hence we now need to get datetime aware
objects from the database as well.

This seems to a bit annoying in (for example SQLite) for the backend.

We use a 3rd-party package to provide us with an `UtcDateTime` type for
SQLAlchemy to fix this.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>