test: conditional e2e discovery
When pytest discovers test files to run it executes the files (or
something like it). This means that imports must be available at that
point in time.

Let's add an import check to `conftest.py` to ignore discovery in the
e2e subdirectory when playwright is not available.

This closes #284.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>